fitbit api

+++ src/main/java/ca/uqtr/fitbit/api/data/Serialization.java
@@ -5,14 +5,14 @@ import lombok.Data;
 @Data
 class Serialization{
     private String dateTime;
-    private int value;
+    private double value;
     private String dataset;
     private int datasetInterval;
 
     public Serialization() {
     }
 
-    public Serialization(String dateTime, int value, String dataset, int datasetInterval) {
+    public Serialization(String dateTime, double value, String dataset, int datasetInterval) {
         this.dateTime = dateTime;
         this.value = value;
         this.dataset = dataset;
